Add proposal for allocation strategies
Allocation strategies is a feature to allow choosing how to allocate LVs
from the VG when using TopoLVM. For example, keep the LV on a single PV
or spread it over multiple.
This proposal suggests extending the current implementation that
supports stripes, to also expose the `alloc` flag of `lvcreate`. It also
proposes a way to ask for a "dedicated" disk so that the LV would
consume exactly one PV.