E2e: Use Recreate strategy for controller
A nodeSelector is used for the controller in the e2e tests. By default
it also has a podAntiAffinity to make controller pods run on different
nodes. This means that during a rolling update, the new pod will be
stuck pending since it must run on the same node as the old.

The Recreate strategy is used instead to fix this. This way the old
controller is first removed before the new pod starts.