example: wait for topolvm controller mutating webhook to become ready
The run target of example Makefile setup topolvm and then deploy sample
pods and pvcs. Deploying pods and pvcs requires that mutating webhook of
the topolvm controller is ready to process requests. Currently we are
waiting for the controller to be the available condition, but the
certificate is also required for webhook to work. This patch reduces
flakiness by waiting for the certificate to become ready.