lvm.go: Utilize lvm JSON and fullreport
Lvm has the option to output JSON.  It also has the ability to grab the
entire state in one command which allows the output to be a consistent state.
Utilize this to reduce the number of fork & exec's of lvm.

Note:
The unit tests use a combination of issuing gRPC calls,
calling directly into the command package, and calling
lvm directly.  When this happens the state of lvm is
incorrect.  Add explicit calls to update the state of
the volume group that has changed.

ref:  "man 8 lvm-fullreport"

Signed-off-by: Tony Asleson <tasleson@redhat.com>