mount: remove new UUID generation
On generating a new UUID(using xfs_admin here), the mount_linux
re-formats the file system; due to which all
the data in the PVC-Clone and PVC-Restore
volume is removed.

To address this issue, we use the `nouuid` mountoption.
Adding 'nouuid' mount option to all XFS mount so that
we will be able to mount a volume and its restored snapshot on
the same node. Without the option, such a mount fails
, because XFS detects that two different volumes with the same
filesystem UUID are being mounted

On the other hand, ext4, by nature, has the nouuid enabled due to which
we dont need to generate a new UUID there too.

Signed-off-by: Yug Gupta <yuggupta27@gmail.com>