webhook: allow PVCs with storageClassName set to ""
A PVC is allowed to set `.spec.storageClassName` equal to "" (empty
string) to bound it to a PV with no storage class. In the current
implementation, for such volumes admission webhook tries to get storage
class with name of "", which fails with the error like "resource name
may not be empty". This error is different from "not found" error and
not well handled, which results in the admission webhook returns an
internal server error.

This patch changes the admission webhook to allow PVCs with
`.spec.storageClassName` set to "" since topolvm does not need to care
about such PVCs.