lvmd: change lv.major and lv.minor to int64
`lvm fullreport` returns negative values for the major and minor number
of LV if the LV is inactive. This commit changes lv.major and lv.minor
to int64 to simplify json.Unmarshal() function and sets major and minor
to 0 for inactive LV when create a LogicalVolume instance. This is safe
since no such device exists.