lvmd: improve arguments of "lvm fullreport"
lvmd executes "lvm fullreport" to get information of VG and LV. This is
more preferable than separate calls to vgs or lvs since the latter can
be inconsistent. The minor downside was its some difficulty in
understanding its options. This commit use more simple but equivalent
option for "lvm fullreport".