Revert "Drop a PVC finalizer to delete pods(#536)"
The PVC finalizer was originally created as a workaround for the following issue.
https://github.com/kubernetes/kubernetes/pull/93457

Because the issue was fixed, the PVC reconciler was once removed by PR #536.
However, it turned out that the PVC finalizer was also useful to
resolve the issue that a pod created from StatefulSet persists in the PENDING state
when PVC is deleted for some reasons like node deletion.
Thus, the reconciler was revived by PR #620.

Signed-off-by: Shinya Hayashi <shinya-hayashi@cybozu.co.jp>
Co-authored-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>