ci: reuse topolvm.img in e2e test matrix
In the e2e test matrix, each job indivisually builds an exactly same
topolvm.img. This PR changes to build topolvm.img in a prerequisite
build job and reuse it in following e2e test jobs.

This change may delay the start of some jobs. However, if there are so
many jobs that not all can be started at the same time, the overall CI
time will be reduced.

Signed-off-by: Daichi Mukai <daichi@mukai.dev>