driver/node: recreate device if owner or device mode is unexpected
When topolvm-node received a NodePublishVolume CSI request, it creates a
device file. It reuse the file if the file already exists and its mode
satisfies some conditions. However, it ignores many parts of the
permission bits and, more over, the owner of the file.

This PR changes to require that the owner and the file type match with
expectation in order to reuse the device file.

Signed-off-by: Daichi Mukai <daichi-mukai@cybozu.co.jp>