Check if a pod of the STS will be re-scheduled to another node
TopoLVM provides node bounded PV. If a pod of a STS uses a TopoLVM PV
and the node where the pod running is deleted, the pod can not be
re-scheduled to another node because the pod uses the PV bounded on the
deleted node. To prevent a such situation, we set a finalizer to nodes
and delete the pod and PVC when a node is deleted. The previous test
only checks the pod and PVC is re-created, now we check re-scheduling a
pod to another node.

Signed-off-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>