Compare topolvm-node metrics to vgs command output directly
Previously, we compared the metrics to node annotation values.  But the
values are for scheduling and do not represent VG size if thinp is used.
Therefore, I compare the metrics to VG output directly.

Also noted that we tested the annotations but they are not intended as
spec, so delete the check.

Signed-off-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>