Sync device-class definitions for both DS LVMd and systemd-service LVMd
Previously, the dc3 device-class was defined for DS LVMd, but not for
the systemd-service LVMd, except on node3. This was because we wanted to
schedule a pod using dc3 on a specific node to simplify the writing of
tests. However, since this can be achieved using nodeSelector, we can
now remove dc3 and synchronize the LVMd configuration.

Signed-off-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>