feat!: Move CRDs to crds folder outside of template
BREAKING CHANGE: Moving to the root crds folder so that crds will not be removed on uninstall. The consequence of this change is CRDs are not upgraded however common CD tools such as argo and flux2 will upgrade the CRDs. It is common for k8s solutions using CRDs to require manual updates when using helm for deployment. Alternative would be to use manifests or olm in the future

Signed-off-by: Ryan Faircloth <ryan@dss-i.com>