feat!: Replace ubuntu with alpine
Alpine is a slightly smaller image this change swaps for a similar configuration and adds nvme-cli and lvm2 so user provided init routines have a likely chance of being able to use the same image to keep pull size down. Result is 284MB vs the current image size of 319MB. A side benefit when the current image is built it is more likely than not a different package will be installed because versions are not easily pined. In this PR I am not pinning versions as this adds a responsibility to the upstream team to update package level locked versions.

Signed-off-by: Ryan Faircloth <ryan@dss-i.com>