docs: remove unnecessary version check
The file `.github/workflows/e2e-k8s-incluster-lvmd.yaml` currently
does not include the steps for tool installation.
These steps were removed in the following PR,
and the versions of those tools are now managed in versions.mk.
https://github.com/topolvm/topolvm/pull/781

Therefore, I'm removing the check for this file
in the tools upgrade procedure.

Signed-off-by: Shinya Hayashi <shinya-hayashi@cybozu.co.jp>