to stable sanity test, hecking the number of pods
Before the change, we checked the number of Pods via daemonset state.
However, when a node is deleted, it does not guarantee the Pod
definition is deleted immediately. Therefore, the previous code could
start a sanity test with the Pod definitions remaining.
With the current code, the sanity test is started with Pod definitions
not remaining.

Signed-off-by: Yuji Ito <llamerada.jp@gmail.com>