Introduce distclean targets
The clean target used to remove not only built files but also downloaded
dependencies. This was undesirable for the development and testing
process, as we need to download necessary binaries many times. To avoid
this, don't remove them by the clean, and do it by the new distclean
target, known as the standard target[1].

[1]: https://www.gnu.org/prep/standards/html_node/Standard-Targets.html

Signed-off-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>