add docker/setup-qemu-action to build-images.yaml
In the build-images process, apt install of libc-bin fails in the
arm64 environment. The following is a response to a similar issue,
and qemu version is expected to be the cause.
https://github.com/docker/buildx/issues/314#issuecomment-1877537571
Looking at the setup-buildx-action documentation, it appears that
qemu is installed using setup-qemu-action, so I will add a similar
process.

Signed-off-by: Yuji Ito <llamerada.jp@gmail.com>