Ignore non-filesystem signatures when re-mounting filesystem PVC
topolvm-node sometimes fails to mount a filesystem PVC that was
previously formatted and mounted to a Pod. This issue appears to stem
from topolvm-node's filesystem detection method, which uses blkid and
may report non-filesystem signatures, such as RAID member signatures
located in the last sector of the block device. Consequently,
topolvm-node interprets this as a filesystem inconsistency error,
resulting in a failure to mount. These should be considered false
positives, as the devices being mounted are intended solely for the
filesystem specified in the PVC spec.

During NodePublishVolume requests, topolvm-node attempts to detect the
filesystem twice. The first detection is performed with its own
implementation, and the second is done by FormatAndMount() provided by
Kubernetes' mount-utils. Currently, both implementations perform
essentially the same function. We can easily modify the former
implementation, but the latter is more challenging to adjust.

To address this issue, this commit modifies topolvm's filesystem
detection method to ignore any usage other than filesystems in its own
implementation. Additionally, we avoid using FormatAndMount() when an
existing filesystem is detected, and instead use Mount(), which does not
check the existing disk format. While this fix does not address false
positives for newly created PVCs, it prioritizes resolving the
re-mounting issue. Newly created PVCs are not yet in use, allowing us
time to address any issues manually.

Signed-off-by: Daichi Mukai <daichi-mukai@cybozu.co.jp>