add option to use LVM commands in containers
This commit adds a new experimental option
`--experimental-run-lvm-commands-in-container` to LVMd.  This option is
completely experimental, and is not expected to become stable in the
near future.

When the flag is enabled, LVM commands are executed inside the LVMd
container itself, eliminating the need for nsenter(2). Consequently, the
commands must be included in the container image, but they no longer
need to exist on the host.

Because this feature is experimental, it is not covered by the e2e
tests. In fact, the current official Docker image does not ship with the
LVM2 utilities, so users who want to try the flag must build a custom
image.

However, I have run all of the tests using a custom configuration, and
all tests passed.  Therefore, I think the quality of this change is
satisfactory.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>