lvmd: support lvm-command-prefix option in lvmd.conf
In DaemonSet mode, LVMd currently executes every LVM command through
/usr/bin/nsenter with hard-coded arguments. This prevents users from:

- Changing the path to nsenter or its options
- Wrapping the command in a logging program or other tool
- Running LVM commands without nsenter (though this is risky)

This commint introduces the lvm-command-prefix option in lvmd.conf to
provide that flexibility. The value is a list of strings that will be
prepended to every command invocation. For example, if it is set to X,
LV creation will be invoked as `X /sbin/lvm lvcreate ...`.

The default value of this option is equivalent to `/usr/bin/nsenter -m
-u -i -n -p -t 1`, so the current behaviour will not break by default.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>